多くの関係者のおかげで2-2が終了いたしました。 将来は、なるべく多くの関係者たちに惜しまれながら死にたいです。
